+++ app/layout.tsx
@@ -3,6 +3,7 @@ import { IBM_Plex_Mono, IBM_Plex_Sans, IBM_Plex_Serif } from 'next/font/google'
 import './globals.css'
 import CookieConsent from '@/components/CookieConsent'
 import ImpersonationBanner from '@/components/ImpersonationBanner'
+import AdminAccessButton from '@/components/AdminAccessButton'
 
 const plexSans = IBM_Plex_Sans({
   subsets: ['latin'],
@@ -134,6 +135,7 @@ export default function RootLayout({
       <body className={`${plexSans.variable} ${plexSerif.variable} ${plexMono.variable}`}>
         {children}
         <CookieConsent />
+        <AdminAccessButton />
         <ImpersonationBanner />
       </body>
     </html>